Fix accessing calls
Some plugins need access to the handler (and rewriter) calls. One
example is the do plugin.

This adds some accessing methods to the callWriter objects, ensures
$calls can be accessed and removes some duplicate code by introducing an
abstract base class.
